Fix PR #39 Copilot review findings: Abs overflow and test handle
Two findings from GitHub Copilot's automated review:

- IsWithinCoalesceWindow's Math.Abs(int) throws OverflowException for
  the one input where the wraparound-corrected signed delta lands
  exactly on int.MinValue (a ~24.8 day gap between two events on the
  same (Hwnd, Kind) pair, expressed as dwmsEventTime's 32-bit modular
  difference). Cast to long before calling Abs — long's range sits
  nowhere near overflowing for an int-derived magnitude, so the
  well-understood wraparound-correction idiom itself is unchanged.
- DwmCloakStateReaderTests used a fixed-looking sentinel HWND
  (0xDEADBEEF) to exercise the invalid-handle path; nint.Zero is
  deterministically never a valid window handle, removing even the
  theoretical (if astronomically unlikely) chance of a false failure.

Added a regression test driving the exact wraparound-boundary input
that used to throw.

+++ src/Bastion.Win32/Coalescer.cs
@@ -444,7 +444,12 @@ internal sealed class Coalescer : BackgroundService
         // as long as the true elapsed time between the two events is under ~24.8 days, which any
         // realistic coalescing window here is, by many orders of magnitude.
         int deltaMs = unchecked((int)(currentDwmsEventTimeMs - lastDwmsEventTimeMs));
-        return Math.Abs(deltaMs) <= _coalesceWindow.TotalMilliseconds;
+
+        // Cast to long before Abs: at the exact wraparound boundary deltaMs can legitimately equal
+        // int.MinValue, and Math.Abs(int) throws OverflowException for that one value specifically
+        // (Copilot review finding on this PR) — Math.Abs(long) has no equivalent edge case here,
+        // since an int-derived magnitude sits nowhere near long.MinValue's.
+        return Math.Abs((long)deltaMs) <= _coalesceWindow.TotalMilliseconds;
     }
 
     private void OnFlushTimerFired(object? state)
